Rename UI components to presentors
