+++ src/presenters/GitInfo.jsx
