+++ src/presenters/IntervalPicker.jsx
