+++ src/presenters/SpinnerOverlay.jsx
